pipeline/requirements.txt
feedparser==6.0.11
python-dotenv==1.0.0
supabase==2.3.0
beautifulsoup4==4.12.3
requests==2.31.0
python-dateutil==2.8.2
